VOInfo: add support for FQAN-like String constructor
Currently, the SpaceManager uses the FQAN class to parse the lines in the 
LinkGroupAuthorizationFile.  This is broken because this file may contain glob patterns and the 
glob wildcard characters ("?" and "*") are not valid in an FQAN.

The current code-base works-by-accident because the FQAN class isn't too fussy about it's input.

This patch adds a constructor to VOInfo so that it can accept a single string (which is either an 
FQAN or an FQAN glob pattern) and extract the group- and role- glob patterns.

The patch removes a static version of the match method as it doesn't add anything.  The two 
enduring usages of the static method are converted to the non-static version of the method.

The patch also adds unit tests for VOInfo to demonstrate correct behaviour.  Some of the tests come 
from the main method in VOInfo.  As these tests are now unit-tests, the patch removes the main 
method.

Patch is for trunk.

Patch: http://rb.dcache.org/r/2069/
Target: trunk
Acked-by: Dmitry