space-manager: use VOInfo objects, rather than FQAN, when reading LGAF
The space-manager uses FQANs to read the LinkGroupAuthorizationFile (LGAF).  This is broken because the LGAF 
may contain glob-patterns (simple pattern-matching with "*" and "?") and neither of the glob wildcard 
characters is valid in an FQAN.

This patch fixes this by creating VOInfo objects.  It is dependent on RB #2069:

    http://rb.dcache.org/r/2069/

The patch also removes the getFqans method from LinkGroupAuthorizationRecord as this method isn't used 
anywhere in dCache.

Target: trunk
Patch: http://rb.dcache.org/r/2071/
Acked-by: Dmitry