unit-tests: factor out serialisation support code
The unit-tests for PnfsId include tests to verify that we can serialise and 
deserialise PnfsId objects.  The code to support these tests is included in the 
PnfsId with the intention to factor it out when we wish to provide similar tests 
for another class.

This patch honours that intention.  It introduces SerializableUtils as a utility 
class for supporting unit-testing of object serialisation.  The PnfsId tests are 
refactored to use SerializableUtils.  The patch also includes some tests for 
SerializableUtils itself.

Patch: http://rb.dcache.org/r/2070/
Target: trunk
Acked-by: Jan