dcap: add storage capacity of string to url2config function
The function url2config (in dcap_url.c) populates the memory pointed to by one of the arguments with 
the configuration option that would result in the supplied dcap url structure.  Currently, this 
function blindly assumes there is sufficient memory to store the necessary configuration string.

This patch adds an additional argument to the url2config function: the storage capacity of the 
string memory allocated.  The ordering of the arguments have been changes to be more in keeping with 
that of snprintf.  The few places where url2config is called have been updated.

This patch does not affect url2config itself, so the assumption of sufficient memory continues.

Patch: http://rb.dcache.org/r/2082/
Acked-by: Tigran
Target: trunk