cells: Avoid PinboardAppender dependency in CellNucleus
PinboardAppender is used to direct log messages to the cell pinboard.

CellNucleus used to register a pinboard statically in PinboardAppender.
This introduced a dependency in CellNucleus on PinboardAppender. This
patch avoids the dependency and the static fields in PinboardAppender
by letting PinboardAppender resolve the cell name through CellNucleus.
Thus PinboardAppender depends on CellNucleus, but CellNucleus does not
depend on PinboardAppender.

This patch is in preparation for replacing the log4j PinboardAppender
with a logback PinboardAppender.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2073/