gplazma: don't hard-code behaviour in strategies, and add ValidationStrategy This is the extention of the patch http://rb.dcache.org/r/1967/ (Paul) , which also adds a ValidationStrategy interface and DoorValidationStrategy implementation. validate method of ValidationStrategy is called in the login method of gPlazma right before the return of the LoginReply. The door specific implementation checks that the attributes and principals needed by typical dCache door are present in the LoginReply.
Here is the original description of http://rb.dcache.org/r/1967/

Patch http://rb.dcache.org/r/1948 included a noticeable deviation from (at least what I believed to be) the gPlazma plugin design: it introduced the concept of a plugin activity throwing an exception (AuthenticationException) and returning a boolean.

The PluginCaller interface, in effect, reintroduced this plugin-returning-a-boolean value as a (class implementing) PluginCaller represents a call to a plugin.  The concrete strategy implementations supply the PluginCaller implementation as an anonymous inner class, so these strategies reintroduce this plugin-returning-a-boolean-value behaviour.

The patch also introduced a number of hard-coded validation requirements that, from what I remember, was going to be implemented as a plugin; for example, the requirement to have a UidPrinciple, at least on GidPrinciple was to be implemented as an GPlazmaAccountPlugin (see 4.14 "validate user" plugin https://srm.fnal.gov/twiki/bin/view/SrmProject/GplazmaProject ).

Constraints on having a home-directory and root-directory should also be represented as a plugin and not hard-coded in the strategies.  We may wish to change these in the future:  adding or removing requirements.  By using plugins and not hard-coding the logic, we make these changes easy.

Because some hard-coded testing has been removed, some nine unit tests had to be updated.  Seven were of the form "empty-configuration should fail" (which shouldn't be enforced in the strategy, so the tests were wrong) and two were the unit-tests were documented as testing broken behaviour (testSufficientPluginFollowedByFailedConfig).

http://rb.dcache.org/r/2079/
Signed-off by: Paul