commons: change FQANPrincipal to use FQAN class
The FQANPrincipal class implements Principal and represents an end-user's FQAN (group) 
membership.

We have a class to represent an FQAN, yet the FQANPrincipal class uses a String to 
represent the underlying FQAN.  This use of a String forces client code to create a new 
FQAN object on receipt of an FQANPrincipal if they wish to use the underlying FQAN's 
structure (group, role, etc).

This patch updates FQANPrincipal to use the FQAN class as underlying storage, rather than a 
String.  It also introduces the public method getFqan to allow calling code to obtain the 
user's FQAN as an FQAN object.

The patch also introduces unit-tests for the FQANPrincipal class (96% code coverage).

FOR RELEASE NOTES:

This patch breaks compatibility between gPlazma, doors, PoolManager and PnfsManager: one 
must undertake a synchronized update of these components.

Patch: http://rb.dcache.org/r/2054/
Target: trunk
Acked-by: Gerd