gplazma: fix XmlResourcePluginRepositoryFactory unit-tests
There are some five unit-tests that are currently failing in Hudson.  These five tests check 
the behaviour of the XmlResourcePluginRepositoryFactory class.

These tests fail because they assume that the XML resource (META-INF/gplazma-plugins.xml) is 
only available from the custom class-loader: Utf8DataClassLoader.  This class provides 
resources for some specified location, independent of JAR files or file-systems.

However, when running the unit-tests from ant, the JAR files (dcache.jar, gplazma.jar, etc) 
are first built and then the tests are run against the compiled JAR files.  The build process 
for gplazma.jar will built it with a META-INF/gplazma-plugins.xml file describing the plugins 
in this JAR file.  As gplazma.jar is in the class-path when running the unit-tests, the 
embedded XML file will be accessible to the system default class-loader.  Having both the 
Utf8DataClassLoader-supplied XML files and the JAR-supplied XML will cause the unit-tests the 
check how many plugins are discovered to fail.

Due to a bug, the XmlResourcePluginRepositoryFactory had previously failed to discover 
resources in the JAR file.  This bug hid the problem with having multiple 
META-INF/gplazma-plugins.xml.  It is only when the bug was fixed by patch

    http://rb.dcache.org/r/2103/

(and committed as r. 14211) that the problem with the unit-tests was revealed.

This patch solves the problem by introducing a new ClassLoader, ResourceBlockingClassLoader, 
that can inhibit the discovery of resources any parent class-loader.  The 
XmlResourcePluginRepositoryFactory tests instantiate ClassLoaders with the following 
hierarchy: Utf8DataClassLoader --> ResourceBlockingClassLoader --> SystemClassLoader.  By 
switching on ResourceBlockingClassLoader's ability to block parent resources, we ensure that 
the test environment contains no unexpected resources and tests that check the number of 
discovered plugins are reliable.

Patch also includes unit-tests for the new ResourceBlockingClassLoader class.

Patch: http://rb.dcache.org/r/2107/
Acked-by: Tigran
Target: trunk