srm: Fix AuthorizationRecord id generation
The persistence manager for AuthorizationRecord relies on id conflicts
being resolved before the AuthorizationRecord is persistet. Previously
the gplazma cell did this in case one authenticated with an x509 chain.
Since we replaced the gplazma cell, this is no longer the case. Hence
the conflict resolution is missing at the moment.

This patch copies the algorithm from the old gPlazma cell to the SRM's
DCacheAuthorization class. Only the SRM and pin manager use persistence
of AuthorizationRecord and PinManager will always get it from the SRM.

I added an additional check to the algorithm to ensure that the
persistet record is identical to the original one.

AFAIK stable releases before 1.9.8 are suffering from the same bug in
case gPlazma is used as a module rather than a cell: Since only the cell
did the conflict resolution, using the module will bypass it.

Target: trunk
Target: 1.9.9 (proposed)
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2133/