srmclient: Fix synchronous/asynchronous check in srmls client
The SRM spec allows many requests to be either processed immediately by
the server (synchronously) or queued (asynchronously). In the first case
the server will typically respond with SRM_SUCCESS in case of success or
an error code in case of failure. In the second case the server will
respond with either SRM_REQUEST_QUEUED or SRM_REQUEST_INPROGRESS:

SRM_REQUEST_QUEUED

§  successful request submission and acceptance. Request token must be returned.

SRM_REQUEST_INPROGRESS

§  Some files are completed, and some files are still on the queue. Details are on the files status.

In that case the client has to poll the server for a result. The request
token is used to link the initial requests and the poll requests
together.

Our srmls client however checked the prensence of a request token to
distinguish between the synchronous and asynchronous reply code. That
assumption is not valid. The spec simply states that if the reply is
asynchronous, then the reply must include a request token. We cannot
from that conclude that the inclusion of a request token means that the
request is processed asynchronously.

This patch replaces the check in the srmclient. Without this patch our
srmclient does not work with dCache 1.9.9. That is due to dCache always
including the request token for reference (e.g. for debugging).

Target: trunk
Target: whatever branch is our production release for the srmclient
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/2138/