space-manager: more fixes in how FQAN is used
The space-manager uses FQAN in a somewhat liberal fashion: to process Strings that might or might 
not be an FQAN.

This patch tidies up these usages so that we check first whether a String is an FQAN before 
attempting to parse it as such.  There should be no change in behaviour of the space-manager as a 
result of the patch.

Patch: http://rb.dcache.org/r/2124/
Acked-by: Thomas
Acked-by: Timur
Target: trunk