admin: don't die if user misconfiguration history file.
RT ticket 5727:

    http://rt.dcache.org/Ticket/Display.html?id=5727

describes how the admin cell will die if the site-admin has configured it with a history file 
that doesn't exist.  Since we package dCache like this, dCache is broken on a fresh install.

The patch fixes this by checking if the configure history file location exists.  If it doesn't, 
the admin cell will attempt to create it.  If this fails or the file exists and isn't writeable 
then an error message is logged and the admin cell will continue to start-up using a per-session 
history for people logging in.

Patch: http://rb.dcache.org/r/2140
Target: trunk
Acked-by: Tanja