pnfsid: add isValid static method
The String constructor of the current PnfsId class implementation will throw an 
IllegalArgumentException if the argument isn't a valid PNFS-ID representation.  However, 
there's currently no way to determine if some arbitrary String is a valid representation.

The dcap door is presented with a string that may be a PNFS-ID representation or a path.  
Since there is no way to determine if the string is a valid PNFS-ID, the door currently 
attempts to create a PnfsId object.  If this results in an IllegalArgumentException then 
the door assumes the string is a path.

Using Exceptions for normal flow-control is considered undesirable at best.

This patch adds the PnfsId.isValid method to allow the dcap door to determine if the 
supplied String is a valid PNFS-ID representation.  A future patch will update the dcap 
door to use this new method.

Patch: http://rb.dcache.org/r/2146/
Target: trunk
Acked-by: Thomas