PnfsManager: Only fold successful messages
PnfsManager supports collapsing processing of identical messages
together and reply to them all in one operation. This is called folding.

The folding code assumes that only successful messages are folded, yet
the code was invoked even for failures. This may lead to wrong replies
being generated. We have seen these at NDGF as NullPointerExceptions in
doors because messages lacked the correct information.

This patch fixes PnfsManager such that it only attempts folding on
successful operations.

Eventually a patch should be written that allows negative replies to be
folded too. Releases before 1.9.9 allowed this, but a recent fix to the
folding code introduced the present problem. This patch represent the
minimal fix for that problem.

Target: trunk
Target: 1.9.9 (proposed)
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/2141/