pool: Generalize target selection of migration module
This patch adds the --exclude-when, -include and -include-when options
to the migration module. The two -when forms accept an expression that
is evaluated for each target pool. Besides its name, the expression may
refer to various dynamic attributes of the pool.

This allows pools to be included/excluded from the migration job based
on their current load or based on how much free/used space they have.

The expression language is implemented by the JEXL library. This library
is added to dCache. This is the first of a series of patches that will
make use of the JEXL library.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/2106/