pool: Let migration module query pool cost for fixed pool list
The migration module can distribute files based on the load of the
destination pool. For pool group and link groups the information was
queried from PoolManager, however for a fixed pool list the load
information was missing.

This patch implements the PoolManagerGetPoolsByNameMessage and updates
the migration module to use it.

RELEASE NOTES: Requires that PoolManager is updated before pools.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/2112/