pool: Refactor PoolListFromPoolManager
PoolListFromPoolManager is a base class of implementations of
RefreshablePoolList used by the migration module. Besides fetching pool
information from PoolManager it supports filtering based on various
criteria.

The filtering doesn't really belong in the class as the primary purpose 
is to fetch the information from PoolManager. Therefore this patch 
refactors the filter code and introduces the new PoolListFilter. This is 
a decorator for the RefreshablePoolList interface.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/2116/