AuthorizationRecord: Fix conversion from Subject
Subjects can be converted to AuthorizationRecord. The conversion
algorithm contains a bug in which it doesn't add a reference from Group
to GroupList. As a result persistence of Group fails. The symptom is
that the AuthorizationRecord in the SRM is lacking its GID.

Target: trunk
Target: 1.9.9 (proposed)
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/2157/