AuthorizationRecord: Fix conversion from Subject
Subjects can be converted to AuthorizationRecord. The conversion
algorithm contains a bug in which it doesn't add a reference from Group
to GroupList. As a result persistence of Group fails. The symptom is
that the AuthorizationRecord in the SRM is lacking its GID.

Patch: http://rb.dcache.org/r/2157/
Acked-by: Paul
Target: trunk (14261)
Target: 1.9.9 (requested)
Request:  http://rt.dcache.org/Ticket/Display.html?id=5755