scripts: fix Solaris compatibility for the daemon script
This patch fixes an issue with the daemon script; namely, there is a variable and a function 
with the same name ("log").  Having a common name for a variable and function is a problem 
for the Solaris shell.

Patch: http://rb.dcache.org/r/1896/
Acked-by: Gerd
Target: trunk (14021)
Target: 1.9.8 (recommended)
Target: 1.9.7 (recommended)
Request: http://rt.dcache.org/Ticket/Display.html?id=5643