srm: Remove log4j initialization
Traditionally the SRM loaded its own log4j configuration. The right
place to load the logging setup is however in the hosting environment
(either as system properties to log4j or in the app server). For dCache
this is handled by the boot loader.

To further remove depedencies on log4j, this patch remove the custom
log4j initialization in the SRM.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Acked-by: Dmitry Litvinsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/2061/