pool: Add source pool constants to exlude and include expression
This patch allows the source pool to be referenced in exclude and
include expression. Thus one can create jobs like

migration copy -exclude-when='target.spaceCost>source.spaceCost' ...

i.e. don't transfer to a pool with a higher space cost than the source
pool.

The source information is like the target information queried from pool
manager. That is because the space cost and cpu cost is computed by
PoolManager rather than the pool. That means the information cannot be
relied on to be 100% up to date; it is refreshed periodically.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/2117/