ftp and pnfsmanager: Check file type on list
PermissionHandlerNameSpaceProvider checks whether the caller is allowed
to list a directory. It does however not check whether the name space
entry is actually a directory (in contrast to a file). Thus if one
attempts to list a file and the permission check fails, then the wrong
error message is generated. The wrong error message is then propagated
to the client.

Debugging this, we also found the error message generated by the FTP
door for NLST and MLSD command when performed on files to be rather
misleading. The error message was updated.

This bug causes compatibility problems with ARC clients and the fix is 
thus proposed for merge into stable branches.

Patch: http://rb.dcache.org/r/1924/
Acked-by: Paul
Target: trunk (14042)
Target: 1.9.8 (proposed)
Target: 1.9.7 (proposed)
Target: 1.9.6 (proposed)
Target: 1.9.5 (proposed)
Request: http://rt.dcache.org/Ticket/Display.html?id=5660