pool: Add lifetime predicates to migration module
The migration module during job initialization builds a list of files to
transfer. It used to run until this list was empty.

This patch adds the options -stop-when=<expr> and -pause-when=<expr> to
influence the lifetime of the job.

Possible use cases are:

- Pause the job when the performance cost of the pool is above a threshold.

- Stop a job when less than a certain amount of files are left.

- Stop a job when all target pools are filled more than 50%. That can be
  achieved with the combination of

    -exclude-when='target.used>target.total/2'

  and

    -stop-when='targets==0'

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/2118/