LinkGroupAuthFile: fix authorisation of linkgroups
Patch #2071

    http://rb.dcache.org/r/2071/

changed the LinkgroupAuthorizationFile parser so it created VOInfo objects, 
rather than FQAN objects, to represent authorisation lines in the file.  This 
is needed as the user may have entered a FQAN-predicate in the 
LinkgroupAuthorizationFile and a FQAN-predicate is NOT a valid FQAN.

This patch requires a previous patch, #2069:

    http://rb.dcache.org/r/2069/

Patch 2069 added a single-string argument constructor to the VOInfo class, 
replacing the work done in LinkgroupAuthorizationFile.

However, Manager further modified the FQAN-predicate.  If a user specifies a 
FQAN-predicate without a role-part (e.g. "/cms") then the role-part is 
altered to the wildcard character ("*"); for example, specifying "/cms" is 
treated as if "/cms/Role=*" has been specified.

The single-String constructor to VOInfo, added in #2069, fails to do this.  
So, in switching Manager to use VOInfo objects created by the new 
single-string constructor, patch #2071 inadvertently altered dCache's 
behaviour.

The code to change empty role-parts to a wildcard is not commented, so there 
is no indication why the code is there.  The modification should not be 
needed and actually hides a bug in gPlazma's X509Utils class.

Patch #2165

  http://rb.dcache.org/r/2165/

addresses the bug in gPlazma.  Once this bug is fixed, the behaviour of 
introducing a "*" is unnecessary and should be removed.

However, in the meantime this patch restores dCache's original behaviour.

Patch: http://rb.dcache.org/r/2158/
Target: trunk
Acked-by: Dmitry