PnfsManager: Resolve regression for ownership of new files
In 1.9.9 several doors delegate the task of choosing the uid, gid and
mode of new files and directories to the PnfsManager. This was done to
both unify the semantics and to allow us to implement the set gid bit on
directories.

The change however causes file ownership rules to change: Previously the
exact semantics did depend on which door was used. Now the uid and gid
is usually taken from the uid and effective gid of the user who created
the name space entry. This happens to be the correct POSIX semantics,
however for shared directories one usually wants at least the gid be
inherited from the parent directory. That is the semantics that can be
achieved by setting the set gid on the directory, however 1.9.9 and
trunk do not yet implement this bit. Even when implemented, it would
require sites to set this bit on all existing directories.

I talked to Paul about this problem and we agree this is to be
considered a regression. The present patch resolves the regression by
introducing the pnfsInheritFileOwnership flag. If set to true, then file
ownership will be inherited from the parent directory. If set to false,
uid and gid are derived from the user creating the file or directory.
Since the latter behaviour is the correct POSIX semantics, the flag
defaults to false. Sites have to explicitly set the flag to true if they
want to maintain the classic behaviour.

Notice that even when set to true, the semantics will have changed for
some doors compared to 1.9.8: Not all doors implemented inheritance for
uid and gid. I do not consider this to be a serious problem: The
benefits of unifying the semantics outweigh the drawbacks.

Target: trunk
Target: 1.9.9 (proposed)
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/2181/