CheckSum: Eliminate diskCacheV111.util.Checksum
The code used two different ways to represent checksums.
diskCacheV111.util.Checksum was a legacy class that was used for both
representing checksums and for computing checksums. The class was not
Serializable. org.dcache.util.Checksum is a newer class that is
Serializable, but it lacked much of the functionality of the other
Checksum class.

This class adds the missing functionality to org.dcache.util.Checksum.
The rest of the code is updated to use this class and the old class is
removed.

Target: trunk
Acked-by: Jan Schaefer <jan.schaefer@desy.de>
Patch: http://rb.dcache.org/r/2170/