Xrootd-mover: Use shared server instance for all movers
Split the functionality of the xrootd-mover into a request handler, that
performs all interactions with the xrootd client and a mover, that performs
interactions with the pool. The request handler will pick the right movers
based on UUID information sent by the door and included as opaque
information by the xrootd client in the redirect. After the open operation,
movers will only be accessed indirectly via the file descriptors by the
pool request handler.

If there are no open channels and no active movers, the mover will shut
down the server blocking the mover port, thus unblocking it. It will be
started again, once there are open client channels again, or a new mover
gets invoked.

Acked-by: Gerd Behrmann <behrmann@ndgf.org>
Target: trunk
Patch: http://rb.dcache.org/r/2096

RELEASE NOTES: Removed the use of a separate port per active Xrootd
mover instance. With the new Xrootd mover design, the mover will only
block a single port on the pool while it is active and no port at all if there
are no active transfers.