Xrootd-pool: Remove filename check in stat
The xrootd-mover can return a simple stat containing the file size 
on an open call. This stat was checking whether the file which was
stated matched the one received from the door - however, this
was unnecessary, because the only operation ever calling stat is
open. Additionally, the check was erroneous because it compared 
local paths with global ones.

Acked-by: Gerd Behrmann <behrmann@ndgf.org>
Target: trunk
Patch: http://rb.dcache.org/r/2207