pool: Fix meta data reconstruction script
The init script supports dumping and restoring the Berkeley DB meta data
of a pool. For this to work the script needs to know the path to the jar
file with the Berkeley DB library.

Last time we updated the library I forgot to update the path used by the
script. This is resolved by this patch.

Target: trunk
Target: 1.9.9 (proposed)
Target: 1.9.8 (proposed)
Target: 1.9.7 (proposed)
Acked-by: Thomas Zangerl <tzangerl@pdc.kth.se>
Patch: http://rb.dcache.org/r/2215/