srm: Skip RQUEUED and READY states for ls
The SRM scheduler supports the RQUEUED and READY states. These are used
for gets and puts to indicate that the TURL is ready to be used by the
client. For requests like ls these states serve no purpose. However if
used, the states are subject to the configured limits for jobs in these
states. For large synchronous ls requests, this may lead to deadlocks as
jobs may be waiting for free slots that can never be provided. The
result is that such jobs would block for ever.

This patch modifies the LsFileRequest such that the job is set to DONE
when it is done processing.

I removed the corresponding configuration options.

Target: trunk
Acked-by: Thomas Zangerl <tzangerl@pdc.kth.se>