PnfsManager: Clean up ChecksumCollection
ChecksumCollection is a utility class used by BasicNameSpaceProvider.

This patch lets ChecksumCollection rely on the ChecksumType and Checksum
classes for most of the processing. At the same time some of the
checksum encoding previously done in BasicNameSpaceProvider (and in the
AttributeChecksumBridge before that) has been moved to
ChecksumCollection.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2187/