BootLoader: Output errors to stderr rather than stdout
Warnings about deprecated properties and the like are logged to the
console. We used to log them to stderr, but with the move to logback
this was accidentally changed to stdout. This patch resolves that
regression.

Without the patch the init script will fail to propagate the errors back
to the user.

Target: trunk
Acked-by: Jan Schaefer <jan.schaefer@desy.de>
Patch: http://rb.dcache.org/r/2219/