srm: store srm file requests in shared memory
During the srm scalability studies it was found out that single node srm which did not use terracotta as network attached memory behaved  differenly from the one that used terracotta. SRM server uses a hashtable of weak references where keys are request and file request ids and values are requests and file requests. The requests and file requests were assumed to be kept from being garbage collected because there were hard references to them from the shared memory hashtable, which is used by rettacotta as a set of roots for clastered objects. I this later table we used to store only requests, as it was assumed that file requests are referenced by requests, so they are not to be garbaged collected. But it turned out that terracotta can remove the objects from memory of the jvm and store them only on the backend server if there are no hard references from this jvm and the objects will be put back into jvm memory only when they are retrieved from the Roots map by calling the get with the appropriate key. So under havy load with limited srm memory the requests and file requests were removed from this jvm memory and the attempts to get the file request back by its id were not successful. 
The led to this exception, which is completely unexpected under non-terracotta configuration:
1 Aug 2010 15:00:21 (jsrm) [] Sheduler(id=put_SRM-cmswn1900) update thread caught an exception !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
org.dcache.srm.SRMInvalidRequestException: jobId = -2144655682 does not correspond to any known job
        at org.dcache.srm.scheduler.Job.getJob(Job.java:357)
        at org.dcache.srm.scheduler.Job.getJob(Job.java:288)
        at org.dcache.srm.scheduler.ModifiableQueue.getGreatestValueObject(ModifiableQueue.java:307)
        at org.dcache.srm.scheduler.Scheduler.updatePriorityThreadQueue(Scheduler.java:498)
        at org.dcache.srm.scheduler.Scheduler.run(Scheduler.java:764)
        at java.lang.Thread.run(Thread.java:619)

This patch maked changes that stores the srm file requests in the shared memory hashtable along with requests. This completely solves the problem.

The relevant paragraph from terracotta documentation is bellow. Nice side effect of this terracotta feature is that SRM can run with lower memory requirements.

Terracotta virtual heap allows arbitrarily large clustered object graphs to fit within a constrained local heap size in client  virtual machines.  The local heap has a "window" on a clustered object  graph.  Portions of a clustered object graph are faulted in and paged  out as needed.  Virtual heap is similar in concept to virtual memory and  is sometimes referred to as "Network Attached Memory."  Clustered objects are lazily loaded from the server as they are accessed by a client  JVM.  This happens by injecting augmented behavior around the GETFIELD  bytecode instruction that checks to see if the object referred to by  that field is currently instantiated on the local heap.  If it isn't on  the local heap yet, the client will request the object from the server,  instantiate the object on the local heap, and ensure that the field  refers to that newly instantiated object.  Conversely, less frequently used objects may be transparently purged  from local heap by Terracotta, subject to the whims of the local JVM  garbage collector.  The amount of clustered object data kept in local  heap is determined at runtime.  The Terracotta Virtual Memory Manager  actively sets references to objects that have fallen out of the cache  to null so that they may become eligible for local garbage collection.   Because clustered objects may be lazily loaded, purged objects will be  transparently retrieved from the server as references to them are  traversed.  For more information on how certain object types and values are loaded into memory, see the Clustered Data Structures Guide.

http://rb.dcache.org/r/2234/
Approved by: Dmitry