ftp: Refactor file transfer related code to use Transfer class
The patch updates the FTP door such that its Transfer class extends the
generic Transfer class already used by the WebDAV and xrootd doors.

In doing so the patch fixes several bugs in the FTP door:

- A deadlock in which the door waits for GFTPTransferStartedMessage even
  though the mover already failed. In this situation the door will
  appear dead as the message thread is blocked by a synchronization in
  the message handler for DoorTransferFinishedMessage. The deadlock will
  break once the client disconnects.

- SocketAdapter leak in the retry loop for downloads. If a pool rejects
  a download, then a new socket adapter is instantiated in each
  iteration. This will cause server sockets to leak.

The patch is relatively large, but it also reduces the size of the FTP
door by 500 lines.

The bugs are present in stable branches too. It is unclear whether there
are easy fixes suitable for stable branches.

The current version still suffers from prolonged blocking of the message
thread when processing DoorTransferFinished: When a proxy is used, then
we wait for up to 5 minutes for the proxy to shut down. During this time
the message thread is blocked. It is however unlikely that it will take
this long for the adapter to shut down. The old code suffered from the
same problem.

Target: trunk