bootloader: Refactor interaction between BootLoader and shell
The BootLoader is responsible for parsing configuration files and for
startin domains. The dCache init script invoked the BootLoader once to
determine the list of domains and once for each domain to determine
configuration parameters for that domain. Finally it was invoked once
for each domain to start the domain.

This patch refactors all this such that the BootLoader only needs to be
invoked once to query configuration parameters for all domains and once
for each domain to start the domain. This simplifies the code and
improves the speed of the init script.

Whereas the old BootLoader would expose configuration parameters by a
list of variable assingments that would be evaluated by the shell script
(an approach that suffered from issues with generating variable names
compatible with the shell), the new BootLoader generates a large shell
function, getProperty, that can provide configuration values for all
properties and all domains. This shell function is evaluated by the
shell script (similar to how we evaluated the variable assignments
before). The benefit is that we can easily expose all configuration
parameters in a single invocation of the boot loader.

A man page for the boot loader was added. The help screen of the boot
loader has been minimized.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/2221/