ftp: Refactor interaction between FTP door and ProxyAdapter
This patch simplifies the ProxyAdapter logic in the FTP door. Previously
the adapter was overloaded as it did two things:

- It managed the server socket for pasv commands
- It acted as proxy between pool and client

The server socket for the pasv command is used for both data transfers
and directory listing.

This overload of the SocketAdapter meant we had a complicated setup with
potentially two SocketAdapter instances that interacted it a rather
complicated way.

This patch:

- Changes the port range logic of the adapters to use PortRange
- Changes the FTP door to manage the server socket for pasv command
- Changes SocketAdapter to adopt the server socket opened by the door

Once SocketAdapter has adobted a server socket it is no longer used by
the FTP door. Before the FTP would have tried to reuse the server socket
for directory listings that happen after a transfer, however it was
doomed to fail: SocketAdapter changes the server socket to non-blocking
mode and that would have failed.

Target: trunk
Acked-by: Thomas Zangerl <tzangerl@pdc.kth.se>
Patch: http://rb.dcache.org/r/2245/