ftp: Simplify performance marker configuration
This patch moves the performance marker configuration to the base class
of the FTP door. This makes performance markers available to all FTP
door dialects rather than just the GSI flavor. This makes sense as the
performance marker functionality is bound to mode E (the grid FTP
extended block mode) and not the encryption of the control channel.

Target: trunk
Acked-by: Thomas Zangerl <tzangerl@pdc.kth.se>
Patch: http://rb.dcache.org/r/2265/