build: fix creation of defaults/classpath.properties file
Patch RB #2295:

    http://rb.dcache.org/r/2295/

which was committed to trunk as r14442, changed how the Java classpath was configured by 
introducing a new auto-generated properties file (classpath.properties) in the defaults 
directory.

The ant's echo task uses a double-dollar to escape literal dollars (i.e., when one wishes to 
prevent property reference expansion).  Unfortunately, current trunk attempts to use a 
backslash for this purpose, which doesn't have the desired effect.  Coincidentally, there is 
no dcache.java.classpath property defined in the ant task, so no expansion takes place and 
the reference remains in the generated classpath.properties file; however, the backslash is 
also present in the output.

This patch replaces the "\$" with "$$" to achieve the desired result.

Patch: http://rb.dcache.org/r/2300/
Target: trunk
Acked-by: Gerd