fix declaration of ourHomeDir property
A previous patch removed all usage of the ourHomeDir property from trunk.  The patch also 
marked the variable as obsolete.

This is unfortunately wrong since (currently) a migrated dCache instance will have a 
dcache.conf derived from a dCacheSetup file, which contains references to ourHomeDir 
property.  The result is that trunk is currently broken for a migrated dCache instance.

This patch fixes this by assigning ourHomeDir the value of the dcache.home property and 
marking ourHomeDir as deprecated.

Patch: http://rb.dcache.org/r/2304
Target: trunk
Acked-by: Jan