scripts: fix loadConfig error message
With a recent shake-up of how scripts are organised, we now have a loadConfig.sh script 
responsible for loading the configuration for a domain.  This script also includes code to 
discover where the java binary is located and, if a suitable java cannot be found, emit an error 
message.

The error message uses the fail function.  Unfortunately this doesn't work.  The fail function 
is defined in utils.sh and loadConfig.sh does source utils.sh (it doesn't know where this file 
is located).  Other scripts that use the fail function have first sourced loadConfig.sh and use 
the domain's configuration to know where utils.sh is located.

This patch fixes the problem by hard-coding the error message.  The structure is what is 
currently produced by printp so the result (from an end-user point-of-view) is the same.

Patch: http://rb.dcache.org/r/2320/
Target: trunk
Acked-by: Jan