scripts: add kill marker when shutting down
When shutting down a domain, we send a SIGTERM and wait for the java process to die.  If the JVM 
has non-daemon threads that have not responded to the corresponding interrupt then the JVM will 
never stop without further "prompting".  So, after ten seconds the dcache script sends a SIGKILL 
to really kill off the JVM.  This can happen if the machine is very busy, but it might also 
indicate a problem somewhere in dCache.

It's ambiguous whether or not the shutdown script needed to send a SIGKILL.

This patch adds an extra identifier so that, if a SIGKILL is needed then a "[K]" is emitted in 
the count sequence.

Patch: http://rb.dcache.org/r/2322/
Acked-by: Jan
Target: trunk